exoplanet_habitability_analyzer/requirements.txt
streamlit>=1.31
pymongo>=4.6
pandas>=2.0
plotly>=5.18
numpy>=1.24
python-dotenv>=1.0
python-dateutil>=2.8